update user model fields

+++ api/mysite/api/models.py
@@ -2,10 +2,10 @@ from django.db import models
 
 # Create your models here.
 class User(models.Model): 
-    id = models.AutoField(primary_key=True)
+    id = models.IntegerField(primary_key=True)
     name = models.CharField(max_length=100)
     age = models.IntegerField()
-    income = models.IntegerField()
+    income = models.FloatField()
     city = models.CharField(max_length=100)    
     
 class Campaign(models.Model):
